Combined Deployment and Start/Stop Event Stream hooks and status streams for better reusablity and maintainability.
